requirements.txt
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.8
einops==0.8.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gymnasium==0.29.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==25.0
pandas==2.2.3
pettingzoo==1.24.3
pillow==10.4.0
proglog==0.1.12
protobuf==6.31.1
pyarrow==18.1.0
pygame==2.3.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyVirtualDisplay==3.0
PyYAML==6.0.2
ray==2.37.0
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
scipy==1.14.1
shellingham==1.5.4
six==1.17.0
sympy==1.13.1
tensorboardX==2.6.4
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typer==0.15.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
